fix: refactor to independant packages
